renamed customOperator to flatMapSync
